# config/openapi.yaml
openapi: 3.0.0
info:
  title: TFT Bible API
  description: API for Teamfight Tactics composition tracking and trait analysis
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.tftbible.com/api/v1
    description: Production server
paths:
...
  /compositions:
    get:
      summary: Get all compositions
      description: Retrieve a paginated list of compositions
      parameters:
        - name: limit
          in: query
          description: Maximum number of results to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: set
          in: query
          description: Filter by TFT set
          required: false
          schema:
            type: string
        - name: tier
          in: query
          description: Filter by composition tier (S, A, B, C, D)
          required: false
          schema:
            type: string
            enum: [S, A, B, C, D]
        - name: category
          in: query
          description: Filter by composition category
          required: false
          schema:
            type: string
        - name: champion
          in: query
          description: Filter by champion name
          required: false
          schema:
            type: string
        - name: difficulty
          in: query
          description: Filter by difficulty level (1-5)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: patch
          in: query
          description: Filter by patch version
          required: false
          schema:
            type: string
        - name: sort_by
          in: query
          description: Sort results by criteria
          required: false
          schema:
            type: string
            enum: [popularity, tier, newest, winrate]
        - name: tags
          in: query
          description: Filter by comma-separated tags
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of compositions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCompositions'
        '400':
          $ref: '#/components/responses/BadRequest'
  /compositions/{id}:
    get:
      summary: Get composition by ID
      description: Retrieve details of a specific composition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Composition ID
      responses:
        '200':
          description: Composition details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Composition'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Composition:
      type: object
      properties:
        id:
          type: string
          example: "comp_12345"
        name:
          type: string
          example: "Ninja Assassin Carry"
        description:
          type: string
          example: "A composition focused on high damage ninja assassins"
        category:
          type: string
          example: "Assassin"
        tags:
          type: array
          items:
            type: string
          example: ["beginner", "aggressive"]
        champions:
          type: array
          items:
            $ref: '#/components/schemas/CompositionChampion'
        augments:
          $ref: '#/components/schemas/CompositionAugments'
        meta:
          $ref: '#/components/schemas/CompositionMeta'
        votes:
          $ref: '#/components/schemas/Votes'
        views:
          type: integer
          example: 1500
        created_at:
          type: string
          format: date-time
          example: "2025-11-20T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-11-21T15:30:00Z"
    CompositionChampion:
      type: object
      properties:
        champion_id:
          type: string
          example: "TFT15_Ahri"
        star_level:
          type: integer
          minimum: 1
          maximum: 3
          example: 2
        items:
          type: array
          items:
            type: string
          example: ["3072", "3089", "3031"]  # Bloodthirster, Rabadon's, Infinity Edge
        position:
          $ref: '#/components/schemas/Position'
        is_core:
          type: boolean
          example: true
        priority:
          type: integer
          example: 1
    Position:
      type: object
      properties:
        x:
          type: integer
          minimum: 0
          maximum: 6
          example: 4
        y:
          type: integer
          minimum: 0
          maximum: 3
          example: 1
    CompositionAugments:
      type: object
      properties:
        preferred:
          type: array
          items:
            type: string
          example: ["big_shot_heart", "giant_slayer"]
        acceptable:
          type: array
          items:
            type: string
          example: ["tactical_resupply", "thrill_of_the_hunt"]
        avoid:
          type: array
          items:
            type: string
          example: ["ascension", "steady_presence"]
    CompositionMeta:
      type: object
      properties:
        tier:
          type: string
          enum: [S, A, B, C, D]
          example: "S"
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
          example: 3
        cost:
          type: string
          example: "Flexible"
        patch:
          type: string
          example: "14.23"
        playstyle:
          type: string
          example: "Aggressive"
        winrate:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.65
        avg_placement:
          type: number
          format: float
          minimum: 1.0
          maximum: 8.0
          example: 2.7
        playrate:
          type: number
          format: float
          example: 0.15
        contest_rate:
          type: number
          format: float
          example: 0.10
    Votes:
      type: object
      properties:
        upvotes:
          type: integer
          example: 120
        downvotes:
          type: integer
          example: 15
...
    PaginatedCompositions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Composition'
        total:
          type: integer
          example: 50
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 3
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters or missing required fields
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Invalid request parameters"
              errors:
                type: array
                items:
                  type: string
                example: ["limit must be between 1 and 100", "invalid trait specified"]
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Resource not found"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT